@@ -17,30 +17,30 @@ var (
 	stream = false
 	req    = [2]api.GenerateRequest{
 		{
-			Model:  "orca-mini",
+			Model:  smol,
 			Prompt: "why is the ocean blue?",
 			Stream: &stream,
-			Options: map[string]interface{}{
+			Options: map[string]any{
 				"seed":        42,
 				"temperature": 0.0,
 			},
 		}, {
-			Model:  "orca-mini",
+			Model:  smol,
 			Prompt: "what is the origin of the us thanksgiving holiday?",
 			Stream: &stream,
-			Options: map[string]interface{}{
+			Options: map[string]any{
 				"seed":        42,
 				"temperature": 0.0,
 			},
 		},
 	}
 	resp = [2][]string{
-		{"sunlight"},
+		{"sunlight", "scattering", "interact"},
 		{"england", "english", "massachusetts", "pilgrims"},
 	}
 )
 
-func TestIntegrationSimpleOrcaMini(t *testing.T) {
+func TestIntegrationSimple(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*120)
 	defer cancel()
 	GenerateTestHelper(ctx, t, req[0], resp[0])
