@@ -1,4 +1,4 @@
 int LLAMA_BUILD_NUMBER = 0;
-char const *LLAMA_COMMIT = "ba1cb19cdd0d92e012e0f6e009e0620f854b6afd";
+char const *LLAMA_COMMIT = "71e90e8813f90097701e62f7fce137d96ddf41e2";
 char const *LLAMA_COMPILER = "";
 char const *LLAMA_BUILD_TARGET = "";
